perf-inject: add --resolve-callchains
The DWARF call-graph mode of perf-record is the most generally
applicable method, but suffers some significant drawbacks compared to
the FP mode:

1. The samples are *extremely* space inefficient. With the default stack
sample size of 8KiB, an example perf record I took when testing this
patch using DWARF call graph information was 33MiB. The equivalent
record storing the raw callchain was 547KiB. If you need to increase the
stack sample size, this problem only gets worse.
2. The samples contain possibly sensitive information in the register
and stack data. While the information we actually want is the raw IPs in
the callchain, that comes along with whatever data happens to have been
stored in registers and the stack at the time. That could include keys,
passwords, personal user information, etc.

Both of these issues can make it difficult to move a perf record from
the place it was created (e.g. a production server) to the place it will
be analyzed (e.g. a developer's local machine).

To resolve this, this patch adds a mode to perf-inject which converts a
--call-graph=dwarf style record into a --call-graph=fp style record by
performing the unwinding and recording the resolved IPs into the
callchain field. You can then use this to postprocess a perf record
in-place before shipping it off for analysis.

Since the kernel-space portion of the sample will be recorded in the
callchain field even when using --call-graph=dwarf, the logic appends
the converted user-space portion to the end of whatever already exists
in the sample's callchain.

Signed-off-by: Steven Fackler <sfackler@gmail.com>